# Conflicts:
#   emulator/src/decoder.h

+++ emulator/src/decoder.h
@@ -2,7 +2,6 @@
 
 #include <stdint.h>
 #include <stdbool.h>
-
 #include "state.h"
 
 /*
@@ -11,56 +10,56 @@
 
 #pragma pack(push, 1)
 typedef struct {
-    uint8_t rd : 5;
+    uint32_t rd : 5;
     uint32_t operand : 18;
-    uint8_t opi : 3;
-    uint8_t : 3;
-    uint8_t opc : 2;
-    uint8_t sf : 1;
+    uint32_t opi : 3;
+    uint32_t : 3;
+    uint32_t opc : 2;
+    uint32_t sf : 1;
 } DPImmInstruction_t;
 #pragma pack(pop)
 
 #pragma pack(push, 1)
 typedef struct {
-    uint8_t rd : 5;
-    uint8_t rn : 5;
-    uint8_t operand : 6;
-    uint8_t rm : 5;
-    uint8_t opr : 4;
-    uint8_t : 3;
-    uint8_t m : 1;
-    uint8_t opc : 2;
-    uint8_t sf : 1;
+    uint32_t rd : 5;
+    uint32_t rn : 5;
+    uint32_t operand : 6;
+    uint32_t rm : 5;
+    uint32_t opr : 4;
+    uint32_t : 3;
+    uint32_t m : 1;
+    uint32_t opc : 2;
+    uint32_t sf : 1;
 } DPRegInstruction_t;
 #pragma pack(pop)
 
 #pragma pack(push, 1)
 typedef struct {
-    uint8_t rt : 5;
-    uint8_t xn : 5;
-    uint16_t offset : 12;
-    uint8_t l : 1;
-    uint8_t : 7;
-    uint8_t sf : 1;
-    uint8_t : 1;
+    uint32_t rt : 5;
+    uint32_t xn : 5;
+    uint32_t offset : 12;
+    uint32_t l : 1;
+    uint32_t : 7;
+    uint32_t sf : 1;
+    uint32_t : 1;
 } SDTInstruction_t;
 #pragma pack(pop)
 
 #pragma pack(push, 1)
 typedef struct {
-    uint8_t rt : 5;
+    uint32_t rt : 5;
     int32_t simm19 : 19;
-    uint8_t : 6;
-    uint8_t sf : 1;
-    uint8_t : 1;
+    uint32_t : 6;
+    uint32_t sf : 1;
+    uint32_t : 1;
 } loadLitInstruction_t;
 #pragma pack(pop)
 
 #pragma pack(push, 1)
 typedef struct {
     uint32_t operand : 26;
-    uint8_t : 4;
-    uint8_t type : 2;
+    uint32_t : 4;
+    uint32_t type : 2;
 } branchInstruction_t;
 #pragma pack(pop)
